add CheckHash and export DecodeHash
This change exports the method DecodeHash and adds a new method CheckHash, which returns
the parameters used to hash the given password. This is useful for real-world use of
argon2id hashing, where you should periodically upgrade the hashing parameters to keep
up with increasing processor speed. In order to do that, you need to get the parameters
a hash was saved with, to compare against your current gold standard.